+++ src/Headers/Utils/WindowUtils.hpp
@@ -7,8 +7,8 @@ namespace System_Utils
 {
     static float scaleX{1};
     static float scaleY{1};
-    static float windowWidth{800};
-    static float windowHeight{500};
+    static int windowWidth{800};
+    static int windowHeight{500};
 
     void updateScales();
     float getCurrentWindowScaleOffset(Vector2 referenceSize);
